+++ packages/editor/js/extensions/hooks/block-settings.js
@@ -39,7 +39,7 @@ import {
 	EditorAdvancedLabelControl,
 } from '../../components';
 import { STORE_NAME } from '../store/constants';
-import { useStoreSelectors } from '../../hooks';
+import { useStoreSelectors, useBlockSideEffectsRestore } from '../../hooks';
 import { sanitizeDefaultAttributes } from './utils';
 import { isBlockTypeExtension, isEnabledExtension } from '../api/utils';
 import { BlockApp, BlockBase, BlockIcon, BlockPortals } from '../components';
@@ -73,6 +73,8 @@ const useSharedBlockSideEffect = (): void => {
 			tabs.style.display = 'block';
 		}
 	}, [selectedBlock]);
+
+	useBlockSideEffectsRestore(selectedBlock);
 };
 
 const EdiBlockWithoutExtensions = ({