+++ packages/editor/js/hooks/index.js
@@ -13,4 +13,5 @@ export { useStoreDispatchers } from './use-store-dispatchers';
 export { useBlockSideEffects } from './use-block-side-effects';
 export { useAdvancedLabelProps } from './use-advanced-label-props';
 export { useDisplayBlockControls } from './use-display-block-controls';
+export { useBlockSideEffectsRestore } from './use-block-side-effects-restore';
 export { useCalculateCurrentAttributes } from './use-calculate-current-attributes';